Add direnv/nvm config files (#5240)
If you use direnv and/or nvm, this helps synchronize the version of Node
you're running in the repo. (Especially helpful right now when flipping
back and forth between main and release-3.0.)

If this causes problems (eg, you do like using direnv but you don't like
this particular way of setting it up) we can make it more configurable.