Make docs peer deps happy
Since we are on npm v7 now we want peer deps to be happy.

The main fix here is downgrading React to v16 because some deps of
Gatsby 2 have peer deps that don't support 17. We also upgrade
`gatsby-theme-apollo-docs` to a version with some fixes (like rolling
back `gatsby-plugin-google-tagmanager` to v2) and we end up pulling in
(via package-lock) the newly published `@apollo/space-kit@5.6.2` which
fixes the React peer deps (in a way that was already fixed in a later
major version).