Renovate: start running against release-3.0
We are in the final days of AS3 development and our focus is no longer
on `main`. Let's let Renovate help us with `release-3.0` instead of
`main`. Note that the `main` renovate.json5 is still what is used for
configuration.

We can remove all the blocked major version bumps because they are not
relevant on AS3.