Require `prettier` formatting again
Prior to 846814ed7 in August 2019, this repository required all code to be
formatted with the tool Prettier. We decided to remove this requirement;
@abernix included a compelling explanation in that commit's message.

The state of prettier in the repository has been somewhat mixed recently. I'm
currently the lead maintainer of Apollo Server and I've kept all the files I
work on most actively (including most of apollo-server-core) prettier-clean when
feasible. In my personal experience I've found that the ability to completely
turn off any part of my brain that thinks about the details of formatting
outweighs the drawbacks described in the above message. I've also found that
many contributors do run prettier accidentally as part of their commits; telling
people to revert changes (which often increase readability!) feels like as much
of a waste of time as forcing folks who may have never encountered the quite
popular Prettier tool to run it. We do use enforced Prettier on several of
Apollo's internal projects and it does seem to work well.

As an experiment, we're going to try turning Prettier enforcement back on. If
this proves to be a hurdle for contributors, we can revert it. I recognize
there's some hubris in ignoring @abernix's well-reasoned plea from two years
ago, and there's a very good chance that in a few months I'll have learned the
hard way that his judgment two years ago was better than my judgment today.

Note that since we've upgraded to Prettier 2 since the last time we enforced
Prettier, so as the comment in the old prettierrc suggests, we no longer need
its requirePragma hack.