Fix code broken by useUnknownInCatchVariables
We use `strict`. In TS 4.4 this implies the new
`useUnknownInCatchVariables` setting, which changes `catch (e)` to mean
`catch(e: unknown)` instead of `catch(e: any)`.

This seems like a good default going forward, but error handling in AS
is pretty complex already, and we don't want to introduce any
unintentional runtime changes here. For example, I'm not super aware if
anyone is depending on being able to throw non-Error objects with fields
that are similar to Error, so adding a bunch of "do something different
if the thrown thing isn't `instanceof Error`" sounds scary.

Instead we take a few approaches:
- In places where we assumed the thrown object was an `Error`, add `as
  Error`.
- In places where we assumed the thrown object was a `GraphQLError`, add
  `as GraphQLError` and an `XXX` comment.
- For `HttpQueryError`, there are some places we already did checks
  based on `error.name`. Move that check into a type predicate
  `isHttpQueryError` instead.
- In some tests and Micro-related files, just use `: any` or `as any`.
  (Micro is not as well documented or understood as other frameworks we
  support.)

Let Renovate treat TypeScript like other packages now that we've
unbroken the build.

Fixes #5660.