Circle: use npm v8
This is the same as v7 except it does not support Node v10 any more.
But we don't either.

While we're at it, upgrade all Node versions to the latest version.

Note that Node v16.11.0 pulls in npm v8, so developers working locally
who rely on the version of Node from npm should upgrade to that version.
However, that version isn't available quite yet in the `cimg/node`
CircleCI docker image we use.